feat(addons): :sparkles: Added cdk Diff class to addons
Added the `Diff` class that manages the configuration and rendering of a diff job.
Implemented the `IDiff` interface to ensure adherence to the diff configuration structure.
Also added the `DiffProps` interface to define configuration properties for initializing a Diff instance.
This enhancement provides a way to generate diffs for specified stack names, with optional customization options.