feat(addons): :sparkles: Added cdk Deploy class.
Implemented the `Deploy` class to manage the configuration and rendering of a deployment job. Created the `DeployProps` interface to define configuration properties for initializing a Deploy instance. Also introduced the `IDeploy` interface to ensure adherence to the deployment configuration structure. This enhancement allows for the deployment of specified stacks with customizable options.