feat(addons): :sparkles: Add cdk DiffDeploy class JobCollection and related interfaces.
Implemented the `DiffDeploy` class to manage the configuration and execution of combined Diff and Deploy operations. Introduced the `DiffDeployOpts` interface to define configuration properties for initializing a DiffDeploy instance. Also introduced the `IDiffDeploy` interface to ensure adherence to the combined diff and deployment configuration structure. This enhancement enables streamlined handling of diff and deployment processes for specified stack names.