feat(addons): :sparkles: Introduce Scripts utility class with pipInstallRequirements method
This commit introduces the Scripts utility class, offering a range of scripting functions. Within this class, the pipInstallRequirements method is available to facilitate the installation of project requirements using pipenv and pip. This installation is conditioned by the presence of either a Pipfile.lock or a requirements.txt file. Tests for the pipInstallRequirements method have also been written to ensure its correctness.